requirements.txt
Django
djangorestframework==3.0.2
